Handle backups from stdin
With this change it is possible to do backups with the openshift stdout
as the source. For this to work properly most of the generic command
section and the output handling had to be rewritten.

Now the outputs of the commands are printed in realtime as they happen.
Also the stdout and stderr of the commands are handled seperately. This
makes detecting errors much easier. To determine if/how many errors
there were the lines from stderr are counted since restic only prints
errors to stderr.