Small racecondition fixed and logging for post
If if `cmd.Wait()` was reached before all output was processed it was
possible to encounter a "File already closed" error. This is due to
`cmd.Wait()` closing the pipes of the command. By moving the output
locks to before `cmd.Wait()` this issue gets resolved because the
output processing is guaranteed to be complete at that point.