Redisign for wrestic
This redisign introduces an API for restic. This will make it easier to
handle the commands and their output. Also done in this redisign is a
new output module which handles errors/webhooks and prom metrics. Lastly
the availability of the restic backend is now tested directly via a S3
client to get rid of the timeout in the snapshot list.

Relates to: APPU-1385