Add .travis.yml
The reaseon for the deploy-config is that in PRs, the master
branch will be built (and pushed) too, which is undesired.
That does not happen with a dedicated deploy block.
Also the 2 deploy providers are necessary due to the bug in travis
that it cannot AND/OR the conditions (some bash syntax error)
Separated build and push stages do not work, as they don't
share the data between.
This config now builds everything, but only pushes dev, master and tags