Ability to restore to the root of a PVC
With this commit wrestic will trim away the first two levels of the
restore path. The restore path contains "/data" + PVC name when backed
up by the k8up operator. So if restoring to the PVC "restore" it would
create this:

restore:/data/PVCName/<data>

With the trim it will cut away "/data/PVCName" and the result will be:

restore:/<data>

So the PVC is a carbon copy of the original and can be mounted and used
without any manual interventions.