Fixing more issues with closed pipes
Added to code to ensure that all std* IO is done before Wait() is called
to ensure that no deadlocks happen.

Also closing the ReadPipe now with io.EOF and filtering that out.
Without the explicit io.EOF a pipe will return io.ErrClosedPipe. Thus we
can differentiate between the intended closing and unintended closing of
the ReadPipe.